+++ data/scripts/entity/merchants/tradingpost.lua
@@ -11,7 +11,7 @@ end
 
 function onPurchaseWarbondsInteraction()
     local entity = Entity()
-    local cw_success = pcall(include, "cosmicwarbridge")
+    local cw_success = true; include("cosmicwarbridge")
     if cw_success and CosmicWarBridge then
         local heat = CosmicWarBridge.getFactionWarHeat(entity.factionIndex) or 0
         if heat < 0.25 then