+++ data/scripts/events/cw_blockade.lua
@@ -60,7 +60,7 @@ function cw_blockade.initialize()
             category = "War Update",
             content = attacker.name .. " forces have established a blockade on the outskirts of sector (" .. x .. ":" .. y .. "). All neutral merchants and civilian vessels are advised to steer clear or risk being fired upon."
         }
-        local cvn_success, cvn = pcall(include, "cosmicvaultnews")
+        local cvn_success, cvn = true, include("cosmicvaultnews")
         if cvn_success and cvn and cvn.publishArticle then
             cvn.publishArticle(article)
         else