+++ data/scripts/server/background/cosmicwarceasefires.lua
@@ -67,7 +67,7 @@ function CosmicWarCeasefires.update(timeStep)
     local cfg = getCfg()
     local rivalryThreshold = cfg.rivalryThreshold or -45000
 
-    local cv_task_success, cv_task = pcall(include, "cosmicvaulttask")
+    local cv_task_success, cv_task = true, include("cosmicvaulttask")
     if cv_task_success and cv_task and cv_task.RunAsync then
         cv_task.RunAsync("CosmicWarCeasefires", function()
             local eased = 0
@@ -98,7 +98,7 @@ function CosmicWarCeasefires.update(timeStep)
                                 local ceasefireChance = cfg.ceasefireChance or 0.25
                                 if rel > rivalryThreshold and random:test(ceasefireChance) then
                                     local gain = random:getInt(2000, 6000)
-                                    local cvf_success, cvf = pcall(include, "cosmicvaultfaction")
+                                    local cvf_success, cvf = true, include("cosmicvaultfaction")
                                     if cvf_success and cvf and cvf.changeRelations then
                                         cvf.changeRelations(a.index, b.index, gain)
                                     else
@@ -151,7 +151,7 @@ function CosmicWarCeasefires.update(timeStep)
                             local ceasefireChance = cfg.ceasefireChance or 0.25
                             if rel > rivalryThreshold and random:test(ceasefireChance) then
                                 local gain = random:getInt(2000, 6000)
-                                local cvf_success, cvf = pcall(include, "cosmicvaultfaction")
+                                local cvf_success, cvf = true, include("cosmicvaultfaction")
                                 if cvf_success and cvf and cvf.changeRelations then
                                     cvf.changeRelations(a.index, b.index, gain)
                                 else